Updated docker-based release process
Read do-release.sh for a thorough description.

Basic idea is that every thing is done inside a Docker container
and Git tags and branches are pushed to a GitHub user repository
(not the official wildfly org repos).
WildFly Full also uses Nexus staged repositories so that we do not need
to release WildFly Core repo *before* we confirm they work with WildFly
full.

At the end of the script, the release manager will have to:
1. push the tag and branch to officila wildfly-core repo
2. release Nexus staging repository.
3. Usual JIRA/CI clean up described at
https://developer.jboss.org/wiki/WildFlyCoreReleaseProcess